refactor(dev): CTL-529 phase 3 — extract broker projection.mjs
Moves the durable-projection layer out of index.mjs into projection.mjs:
interest persistence (saveInterests / loadPersistedInterests /
migrateLegacyInterestsFile), the broker.state.json builder/writer
(buildBrokerState / writeBrokerStateFile / getBrokerStateFilePath /
persistBrokerState), and the worker-state shadow writers
(getProjectedWorkerStatePath / writeProjectedWorkerState). projection.mjs
imports config + state + broker-state.mjs only — never router or tailer —
so it sits below the router in the dependency DAG.

Plan adaptation: handleWorkerStateChanged stays with the other event
handlers (→ router.mjs in phase 4) rather than moving to projection.mjs
as the plan's section map implied. It reads events via the canonical-aware
getEventPayload/getEventOrchestrator helpers, which belong to the router;
housing the handler in projection would force a projection→router import
and break the acyclic DAG. The router→projection edge already exists, so
the handler reaches the projected-file path/writer helpers cleanly. Same
cycle-safety rationale the plan applied to getEventLogPath.

index.mjs re-exports the 7 public projection symbols; barrel contract and
worker-state / phase-lifecycle persistence tests stay green.

Broker suite green: 273 pass / 0 fail across 7 files.
